# services/gateway/src/openapi.yaml
openapi: 3.0.3
info:
  title: Dental Clinic Appointment API
  version: "1.0.0"
  description: >
    Public API surface for the dental appointment booking system, served
    behind the gateway. Internally this fans out to independent
    microservices (patient, dentist, appointment, reminder, notification,
    report) — but every route below is reached the same way it always was,
    as a single relative `/api/*` origin.
servers:
  - url: /api
    description: Through the gateway (relative — works from any host it's deployed behind)
  - url: http://localhost:4000/api
    description: Local dev
...
paths:
...
  /appointments:
    get:
      tags: [Appointments]
      summary: List appointments
      parameters:
        - name: status
          in: query
          description: Comma-separated statuses
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: dentistId
          in: query
          schema: { type: integer }
        - name: patientId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Appointment" }
    post:
      tags: [Appointments]
      summary: Book an appointment
      description: >
        Validates patient_id/dentist_id against patient-/dentist-service
        (fails closed with 503 if either is unreachable), checks for a
        conflicting slot (409), schedules reminders, and fires a
        confirmation email.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingInput" }
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "400":
          description: Invalid patient_id/dentist_id/time, or time in the past
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: that slot is already booked (double-booking prevented)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: unable to validate patient/dentist right now, please retry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /appointments/{id}:
    get:
      tags: [Appointments]
      summary: Get one appointment
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "404":
          description: appointment not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /appointments/{id}/reschedule:
    patch:
      tags: [Appointments]
      summary: Move a booked appointment to a new time
      description: Re-issues reminders for the new time.
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RescheduleInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "400":
          description: Not booked, invalid time, or time in the past
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: appointment not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: that slot is already booked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: unable to validate dentist right now, please retry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /appointments/{id}/cancel:
    patch:
      tags: [Appointments]
      summary: Cancel an appointment
      description: Voids pending reminders and sends a cancellation email.
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "400":
          description: already cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: appointment not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /appointments/{id}/status:
    patch:
      tags: [Appointments]
      summary: Set appointment status
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [booked, completed, no_show]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "400":
          description: invalid status value
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: appointment not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /appointments/{id}/follow-up:
    post:
      tags: [Appointments]
      summary: Book a follow-up linked to a completed/previous appointment
      parameters:
        - $ref: "#/components/parameters/AppointmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FollowUpInput" }
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "400":
          description: Invalid dentist_id/time, or time in the past
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: appointment not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: that slot is already booked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: unable to validate patient/dentist right now, please retry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  parameters:
...
    AppointmentId:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Appointment:
      type: object
      properties:
        id: { type: integer }
        patient_id: { type: integer }
        dentist_id: { type: integer }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        status: { type: string, enum: [booked, completed, cancelled, no_show] }
        reason: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        follow_up_of: { type: integer, nullable: true }
        cancel_reason: { type: string, nullable: true }
        cancelled_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        patient_name: { type: string, nullable: true, description: Composed from patient-service }
        patient_email: { type: string, nullable: true, description: Composed from patient-service }
        patient_phone: { type: string, nullable: true, description: Composed from patient-service }
        dentist_name: { type: string, nullable: true, description: Composed from dentist-service }
    BookingInput:
      type: object
      required: [patient_id, dentist_id, start]
      properties:
        patient_id: { type: integer }
        dentist_id: { type: integer }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time, description: "Optional — defaults from duration_minutes / the dentist's slot length" }
        duration_minutes: { type: integer }
        reason: { type: string }
        notes: { type: string }
    RescheduleInput:
      type: object
      required: [start]
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        duration_minutes: { type: integer }
    FollowUpInput:
      type: object
      required: [start]
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        duration_minutes: { type: integer }
        dentist_id: { type: integer, description: Defaults to the parent appointment's dentist }
        reason: { type: string }
        notes: { type: string }
...